feat: move subscription query to query string
