+++ src/domain/onboarding/api/create-profile/create-profile.action.ts
@@ -8,7 +8,7 @@ import { EApiEndpoint } from '@enums';
 
 import { STARTUP_CLIENT_TYPE_ID } from '@constants';
 
-import { IDesktopOnboardingMachineContext } from '../../desktop-onboarding/machines';
+import { IDesktopOnboardingMachineContext } from '../../organisms/desktop-onboarding/machines';
 
 export const createAccountAction = async (data: IDesktopOnboardingMachineContext) => {
   const isStartupPath = data.stepTwoData.clientTypeId === STARTUP_CLIENT_TYPE_ID;
