+++ src/domain/profile/api/index.ts
@@ -1,0 +1,1 @@
+export * from './subscription/subscription.action';
