+++ src/domain/profile/api/subscription/subscription.action.ts
@@ -1,0 +1,6 @@
+import { nextAxiosInstance } from '@infrastructure';
+
+import { EApiEndpoint } from '@enums';
+
+export const subscriptionAction = async (planId: string) =>
+  nextAxiosInstance.get(`${EApiEndpoint.SUBSCRIPTION}/${planId}`);
